working on iisuuess

+++ work-smart-backend/aircore/src/main/java/com/aircore/entity/LeaveRequest.java
@@ -41,7 +41,7 @@ public class LeaveRequest {
 	@Temporal(TemporalType.TIMESTAMP)
 	private Date createDate;
 	private Date updatedDate;
-	private Integer appliedDays;
+	private Double appliedDays;
 	@PrePersist
 	protected void onCreate() {
 		this.createDate = new Date();
